requirements.txt
Django
djangorestframework
celery
django-celery-results
redis
psycopg2-binary
python-dotenv
